// lib/api/schema.d.ts
export interface components {
    schemas: {
        CreateBoardRequest: {
            /** @description The question for the board */
            question: string;
            /**
             * Format: int64
             * @description Time-to-live in seconds
             */
            ttl: number;
        };
        AddWordRequest: {
            /** @description The word to add to the board */
            word: string;
        };
        BoardResponse: {
            /** @description The unique identifier of the board */
            id: string;
            /** @description The board's question */
            question: string;
            /**
             * Format: date-time
             * @description When the board expires
             */
            expiresAt: string;
            /** @description A map of words and their counts */
            words: {
                [key: string]: number;
            };
        };
        ErrorResponse: {
            /** @description A description of the error */
            message: string;
            /** @description Optional detail or root cause of the error */
            cause?: string | null;
        };
        BoardEvent: {
            /** @description The type of the event (word-added or board-created) */
            eventType: string;
            /** @description The ID of the board the event relates to */
            boardId: string;
        };
        WordAddedEvent: {
            eventType: "WordAddedEvent";
        } & (Omit<components["schemas"]["BoardEvent"], "eventType"> & {
            word: components["schemas"]["Word"];
        });
        BoardCreatedEvent: {
            eventType: "BoardCreatedEvent";
        } & Omit<components["schemas"]["BoardEvent"], "eventType">;
        Word: {
            /** @description The word text */
            value: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}